requirements.txt
streamlit-pandas == 0.0.9
pandas === 1.5.3
pycryptodome == 3.19.0
st-pages == 0.4.5